junker shop fixes

+++ dumb-compile/compile-and-run.sh
@@ -7,6 +7,12 @@ if [ ! -d $build_dir ]; then
 fi
 
 cd $build_dir
+
+if [ $? -ne 0 ]; then
+  echo "Cannot cd into build directory!"
+  exit 1;
+fi
+
 rm -rf *
 cmake -D CMAKE_BUILD_TYPE=Release ../../ || cmake3 -D CMAKE_BUILD_TYPE=Release ../../
 make -j4
